# static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RosterIQ API",
    "description": "AI-powered predictive rostering for Australian hospitality venues. Manage venues, employees, rosters, forecasts, and compliance with intelligent optimization.",
    "version": "1.0.0",
    "contact": {
      "name": "RosterIQ Support",
      "url": "https://rosteriq.io",
      "email": "support@rosteriq.io"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://rosteriq.io/license"
    }
  },
  "servers": [
    {
      "url": "https://api.rosteriq.io",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service authentication"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "example": "user_123abc"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "manager@venue.com.au"
          },
          "name": {
            "type": "string",
            "example": "Jane Smith"
          },
          "organisation": {
            "type": "string",
            "example": "org_456def"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-25T10:30:00Z"
          }
        },
        "required": ["user_id", "email", "name"]
      },
...
    }
  },
  "paths": {
...
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current user profile",
        "operationId": "get_current_user",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "Current user details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": ["Auth"],
        "summary": "Update user profile",
        "operationId": "update_user",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
...
  }
}